# kv: don't perform async intent resolution on 1PC with point lock spans